requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
neo4j==5.27.0
python-dotenv==1.0.1
requests==2.32.3
eventlet==0.37.0
